[risk] Rename types to domain
