+++ projects/ores.risk/domain/currency_config.hpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2024 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -17,16 +17,28 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#include <ostream>
-#include <rfl.hpp>
-#include <rfl/json.hpp>
-#include "ores.risk/types/currency_config.hpp"
+#ifndef ORES_RISK_DOMAIN_CURRENCY_CONFIG_HPP
+#define ORES_RISK_DOMAIN_CURRENCY_CONFIG_HPP
 
-namespace ores::risk::types {
+#if defined(_MSC_VER) && (_MSC_VER >= 1200)
+#pragma once
+#endif
 
-std::ostream& operator<<(std::ostream& s, const currency_config& v) {
-    rfl::json::write(v, s);
-    return(s);
+#include <iosfwd>
+#include <vector>
+#include "ores.risk/domain/currency.hpp"
+
+namespace ores::risk::domain {
+
+/**
+ * @brief Represents a set of ORE currencies.
+ */
+struct currency_config {
+    std::vector<currency> currencies;
+};
+
+std::ostream& operator<<(std::ostream& s, const currency_config& v);
+
 }
 
-}
+#endif
